Update more tests to handle new Participant equals method
assertEquals may not be used in these tests due to the new equals method
in Participant. isSameParticipant needs to be used instead. Therefore,
instead of using assertEquals on two models or two address books,
assertEquals can be used on everything (user prefs, event list) except
the participant list, in which case a for loop and isSameParticipant
is used to check equality.

Reasons for this is that some of these tests require the creation of
two distinct Participant objects, so the equals method will return false
for comparing them.